LayoutTests/fast/shadow-dom/direct-focus-pseudo-does-not-match-in-author-stylesheet-expected.txt

This tests that -internal-direct-focus is not recognized in author stylesheets

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS defaultFocus.matches(":focus") is true
PASS defaultFocus.matches(":-internal-direct-focus") threw exception SyntaxError: ':-internal-direct-focus' is not a valid selector..
PASS inputInShadowRoot.matches(":focus") is true
PASS inputInShadowRoot.matches(":-internal-direct-focus") threw exception SyntaxError: ':-internal-direct-focus' is not a valid selector..
PASS successfullyParsed is true

TEST COMPLETE
